Move session into separated module
